Implemented SqlTraceWorkloadListener

+++ WorkloadTools/Consumer/Analysis/WorkloadAnalyzer.cs
@@ -165,7 +165,7 @@ namespace WorkloadTools.Consumer.Analysis
 
             DataRow row = rawData.NewRow();
 
-            string normSql = normalizer.NormalizeSqlText(evt.Text, evt.Type.ToString(), evt.SPID);
+            string normSql = normalizer.NormalizeSqlText(evt.Text, evt.Type.ToString(), (int)evt.SPID);
             long hash = normalizer.GetHashCode(normSql);
 
             if (!normalizedQueries.ContainsKey(hash))
